+++ FrontEnd/android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.frontend",
+    package = "com.mercado",
 )
 
 android_resource(
     name = "res",
-    package = "com.frontend",
+    package = "com.mercado",
     res = "src/main/res",
 )
