+++ FrontEnd/android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'FrontEnd'
+rootProject.name = 'Mercado'
 
 include ':app'
