+++ FrontEnd/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "FrontEnd",
-  "displayName": "FrontEnd"
+  "name": "Mercado",
+  "displayName": "Mercado"
 }
