+++ FrontEnd/ios/Mercado.xcodeproj/xcshareddata/xcschemes/Mercado.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "FrontEnd.app"
-               BlueprintName = "FrontEnd"
-               ReferencedContainer = "container:FrontEnd.xcodeproj">
+               BuildableName = "Mercado.app"
+               BlueprintName = "Mercado"
+               ReferencedContainer = "container:Mercado.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "FrontEndTests.xctest"
-               BlueprintName = "FrontEndTests"
-               ReferencedContainer = "container:FrontEnd.xcodeproj">
+               BuildableName = "MercadoTests.xctest"
+               BlueprintName = "MercadoTests"
+               ReferencedContainer = "container:Mercado.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "FrontEndTests.xctest"
-               BlueprintName = "FrontEndTests"
-               ReferencedContainer = "container:FrontEnd.xcodeproj">
+               BuildableName = "MercadoTests.xctest"
+               BlueprintName = "MercadoTests"
+               ReferencedContainer = "container:Mercado.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "FrontEnd.app"
-            BlueprintName = "FrontEnd"
-            ReferencedContainer = "container:FrontEnd.xcodeproj">
+            BuildableName = "Mercado.app"
+            BlueprintName = "Mercado"
+            ReferencedContainer = "container:Mercado.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "FrontEnd.app"
-            BlueprintName = "FrontEnd"
-            ReferencedContainer = "container:FrontEnd.xcodeproj">
+            BuildableName = "Mercado.app"
+            BlueprintName = "Mercado"
+            ReferencedContainer = "container:Mercado.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "FrontEnd.app"
-            BlueprintName = "FrontEnd"
-            ReferencedContainer = "container:FrontEnd.xcodeproj">
+            BuildableName = "Mercado.app"
+            BlueprintName = "Mercado"
+            ReferencedContainer = "container:Mercado.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
