+++ FrontEnd/ios/Mercado/Images.xcassets/AppIcon.appiconset/Contents.json
