+++ FrontEnd/ios/MercadoTests/MercadoTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface FrontEndTests : XCTestCase
+@interface MercadoTests : XCTestCase
 
 @end
 
-@implementation FrontEndTests
+@implementation MercadoTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
